Major Update - Remove unneeded `data` folder at root of project - Deprecate API_Utilities class (kept only for the shaEncoder function) - Begin migration to better organization
