--- ESP/data/ascii.txt
@@ -1,60 +1,0 @@
-                                                                      : === WELCOME === TO === :
-                        ===========================================================================================================================
-                                                        ██████╗ ██████╗ ███████╗███╗   ██╗██╗██████╗ ██╗███████╗
-                                                       ██╔═══██╗██╔══██╗██╔════╝████╗  ██║██║██╔══██╗██║██╔════╝
-                                                       ██║   ██║██████╔╝█████╗  ██╔██╗ ██║██║██████╔╝██║███████╗
-                                                       ██║   ██║██╔═══╝ ██╔══╝  ██║╚██╗██║██║██╔══██╗██║╚════██║
-                                                       ╚██████╔╝██║     ███████╗██║ ╚████║██║██║  ██║██║███████║
-                                                        ╚═════╝ ╚═╝     ╚══════╝╚═╝  ╚═══╝╚═╝╚═╝  ╚═╝╚═╝╚══════╝
-
-                                                                                ██████████████                            
-                                                                          ██▓▓▓▓▓▓▒▒▒▒▒▒▒▒░░▒▒▓▓▓▓██                      
-                                                                      ████▓▓░░░░▒▒░░░░░░▒▒░░░░░░▒▒░░████                  
-                                                                  ██▓▓▒▒▓▓▓▓▒▒▒▒░░░░░░▒▒░░▒▒░░░░░░▒▒░░▒▒▓▓▓▓              
-                                                                ██▓▓▒▒▒▒▒▒▒▒░░▒▒░░░░░░░░░░░░▒▒░░░░▒▒░░░░▒▒░░██            
-                                                              ██▓▓▓▓░░░░▒▒░░░░▒▒▒▒░░░░░░░░░░▒▒░░  ░░░░░░░░▒▒░░██          
-                                                            ██▓▓▓▓▓▓▓▓▓▓░░░░░░▒▒░░░░░░░░░░░░░░      ░░░░░░░░░░░░██        
-                                                          ██▓▓▒▒▒▒▒▒▒▒▒▒▒▒░░░░░░▒▒░░░░░░░░░░░░  ░░            ▒▒▒▒██      
-                                                          ██▓▓▒▒▒▒▒▒▒▒░░░░▒▒░░░░░░░░░░░░░░░░░░                ░░░░██      
-                                                        ▓▓▓▓▒▒▒▒▒▒▒▒▒▒░░░░░░▒▒░░░░░░░░░░░░░░░░░░▓▓▓▓▓▓▓▓▓▓    ░░  ▒▒▓▓    
-                                                        ██▒▒▒▒▒▒▒▒▒▒░░░░░░░░░░░░░░░░░░░░░░░░▓▓▓▓▒▒▒▒▒▒▒▒▓▓▓▓▓▓    ░░██    
-                                                      ▓▓▒▒▒▒▒▒▒▒▒▒░░░░░░░░░░░░░░░░░░░░    ▓▓▒▒▒▒▒▒▒▒░░░░▒▒▒▒▓▓▒▒  ░░▒▒▓▓  
-                                                      ██▓▓▒▒░░░░░░▒▒▒▒▒▒░░░░░░░░░░░░░░░░▓▓▒▒▒▒▒▒▒▒▒▒▒▒▒▒░░░░▒▒▓▓▓▓  ░░██  
-                                                      ██▓▓▓▓▒▒▒▒▒▒▒▒░░░░▒▒░░░░░░░░░░░░  ▒▒▒▒▒▒▒▒▒▒▒▒████▓▓░░░░▒▒▓▓  ░░██  
-                                                    ██▓▓▒▒▒▒▒▒▓▓▒▒▓▓░░░░░░░░░░░░░░░░░░▓▓▒▒▒▒░░▒▒▒▒████  ▒▒██░░▒▒▓▓▓▓  ░░██
-                                                    ██▒▒▒▒▒▒▒▒▒▒▒▒▒▒░░░░░░░░░░░░░░░░░░▓▓░░▒▒▒▒▒▒██████▒▒  ▓▓▓▓░░▓▓▓▓  ░░██
-                                                    ██▓▓▒▒▒▒░░▒▒▒▒▒▒░░░░░░░░░░░░░░░░░░▓▓▒▒░░▒▒░░████████▓▓  ██▓▓▒▒▓▓  ░░██
-                                                    ██▓▓▓▓▓▓▓▓▒▒▒▒▒▒░░░░░░▒▒░░░░░░░░░░▒▒░░▒▒░░▒▒██████████▒▒██▒▒▒▒▓▓  ░░██
-                                                    ██▓▓▒▒▒▒▒▒░░░░░░░░▒▒▒▒░░░░░░░░░░░░▒▒▒▒░░░░▒▒██▒▒██████  ██▒▒▒▒▓▓  ░░██
-                                                    ██▒▒▒▒▒▒░░▓▓▓▓▓▓▒▒░░░░░░░░░░░░░░░░▒▒░░  ▒▒░░██  ▒▒████▒▒██▒▒▒▒▓▓  ░░██
-                                                    ██▓▓▓▓▓▓▓▓░░▒▒▒▒░░▒▒░░░░░░░░░░░░░░▒▒▒▒░░░░░░▒▒██  ██████▒▒▒▒▒▒▓▓░░░░██
-                                                  ██▓▓██▓▓▒▒▒▒▓▓░░░░▒▒░░░░░░░░░░░░░░░░░░▒▒▒▒░░▒▒░░▒▒██████▒▒▒▒▒▒▓▓  ░░██  
-                                                  ██▓▓██▒▒▒▒▒▒▒▒▓▓▓▓▒▒░░░░░░░░░░░░░░░░░░▒▒░░▒▒░░▒▒░░▒▒░░▒▒▒▒▒▒▒▒▓▓░░░░██  
-                                                ██▒▒▓▓██▓▓▒▒▒▒▒▒▒▒▒▒▒▒░░░░░░░░░░░░░░░░░░░░▒▒░░▒▒░░░░░░▒▒▒▒▒▒▒▒▓▓  ░░░░██  
-                                                ██▒▒▒▒▓▓██▓▓▓▓▒▒▒▒▒▒▒▒░░░░░░░░▒▒░░░░░░░░░░░░▒▒▒▒▒▒░░▒▒▒▒▒▒▓▓▓▓  ░░░░██    
-                                              ▓▓▓▓▒▒▒▒▓▓██▓▓▒▒▒▒▒▒░░░░▒▒▒▒░░▒▒░░░░░░░░░░░░░░░░░░▒▒▒▒▓▓▓▓▓▓░░░░░░░░▒▒██    
-                                              ██▒▒▒▒▓▓▒▒▓▓██▒▒▒▒▒▒▒▒▓▓▒▒▒▒▒▒░░░░░░░░░░░░░░░░░░░░░░░░░░░░  ░░  ░░░░██      
-                                            ██▒▒▒▒▓▓░░▒▒▒▒██▓▓▓▓▓▓▒▒▒▒▒▒▒▒▓▓▓▓░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░▒▒░░▒▒██      
-                                            ██▒▒▒▒▓▓▒▒░░▓▓▒▒██▓▓▒▒▒▒▒▒▒▒░░▓▓▒▒▓▓▒▒░░░░▒▒░░░░░░░░░░░░░░░░░░░░░░▓▓██        
-                                            ██▒▒▒▒▒▒░░▒▒▒▒▒▒▓▓██▓▓▒▒▒▒▒▒▓▓▒▒▒▒▒▒▒▒▒▒▓▓░░░░░░░░░░▒▒░░░░▒▒░░▒▒▒▒██          
-                                            ██▒▒▒▒░░▒▒░░▒▒████  ██▓▓▓▓▓▓▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒░░▓▓▒▒▒▒▒▒▓▓▒▒██            
-                                            ██▒▒▓▓░░▒▒░░▓▓        ████▓▓▒▒▓▓▒▒▒▒▒▒░░▒▒▒▒▒▒▒▒▓▓▒▒▒▒▒▒▒▒▒▒██▓▓              
-                                            ██▒▒▒▒▒▒▒▒  ██            ████▓▓▒▒▓▓▓▓▓▓▒▒▒▒▒▒▒▒▓▓▒▒▒▒▒▒████                  
-                                              ██▒▒▒▒▒▒▒▒░░██              ██████▓▓▒▒▒▒▒▒▒▒▓▓▓▓██████                      
-                                              ██▒▒▒▒▒▒▒▒░░  ████                ██████████████                            
-                                                ██▒▒▒▒▒▒▒▒░░  ░░████                                                      
-                                                  ████▒▒▒▒▒▒░░░░  ░░████                                                  
-                                                      ████▒▒▒▒▒▒░░░░  ░░██                                                
-                                                          ████▒▒▒▒▒▒░░  ░░██                                              
-                                                              ██▓▓▒▒▒▒░░  ▒▒▓▓                                            
-                                                                  ████▒▒░░  ▒▒██                                          
-                                                                      ▓▓▒▒░░░░██                                          
-                                                                        ██░░  ██                                          
-                                                                  ▓▓██  ██░░░░██                                          
-                                                                ██░░██  ██░░░░██                                          
-                                                              ██░░██    ██░░▒▒██                                          
-                                                              ██░░▒▒████░░▒▒██                                            
-                                                                ▓▓▒▒▒▒▒▒▒▒▓▓                                              
-                                                                  ████████                      
-                                                          
-                        ============================================================================================================================
