+++ ESP/lib/src/network/api/utilities/apiUtilities.cpp
@@ -1,28 +1,10 @@
 #include "apiUtilities.hpp"
 
-//! These have to be called before the constructor of the class because they are static
-//! C++ 11 does not have inline variables, sadly. So we have to do this.
-const char *API_Utilities::MIMETYPE_HTML{"text/html"};
-// const char *BaseAPI::MIMETYPE_CSS{"text/css"};
-// const char *BaseAPI::MIMETYPE_JS{"application/javascript"};
-// const char *BaseAPI::MIMETYPE_PNG{"image/png"};
-// const char *BaseAPI::MIMETYPE_JPG{"image/jpeg"};
-// const char *BaseAPI::MIMETYPE_ICO{"image/x-icon"};
-const char *API_Utilities::MIMETYPE_JSON{"application/json"};
-
 //*********************************************************************************************
 //!                                     API Utilities
 //*********************************************************************************************
 
-API_Utilities::API_Utilities(int CONTROL_PORT,
-                             ProjectConfig *projectConfig,
-                             CameraHandler *camera,
-                             StateManager<WiFiState_e> *WiFiStateManager,
-                             const std::string &api_url) : server(new AsyncWebServer(CONTROL_PORT)),
-                                                           projectConfig(projectConfig),
-                                                           WiFiStateManager(WiFiStateManager),
-                                                           camera(camera),
-                                                           api_url(api_url) {}
+API_Utilities::API_Utilities() {};
 
 API_Utilities::~API_Utilities() {}
 
@@ -53,19 +35,4 @@ std::string API_Utilities::shaEncoder(std::string data)
 		hash_string += hex;
 	}
 	return hash_string;
-}
-
-void API_Utilities::notFound(AsyncWebServerRequest *request) const
-{
-	if (_networkMethodsMap.find(request->method()) != _networkMethodsMap.end())
-	{
-		log_i("%s: http://%s%s/\n", _networkMethodsMap.at(request->method()).c_str(), request->host().c_str(), request->url().c_str());
-		char buffer[100];
-		snprintf(buffer, sizeof(buffer), "Request %s Not found: %s", _networkMethodsMap.at(request->method()).c_str(), request->url().c_str());
-		request->send(404, "text/plain", buffer);
-	}
-	else
-	{
-		request->send(404, "text/plain", "Request Not found using unknown method");
-	}
 }
